reafctor: changes for ts using
